backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
torch
transformers
safetensors
numpy
python-multipart
requests
python-dotenv
bcrypt
python-jose[cryptography]
psycopg2-binary
